+++ docs/make_local.jl
@@ -12,6 +12,7 @@ makedocs(
         "Home" => "index.md",
         "Manual" => "manual.md",
         "Use Cases" => "use_cases.md",
+        "Best Practice (Please Read)" => "best_practice.md",
         "API" => "API.md",
         "Change Log" => "changelog.md"
     ]